BE/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# GCP Cloud Storage (instead of S3)
google-cloud-storage==2.19.0

# Auth (for later phases)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Utilities
httpx==0.28.0
email-validator==2.2.0

# CORS
starlette==0.41.3
